requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
easyocr==1.7.2
fastapi==0.128.3
h11==0.16.0
idna==3.11
numpy==2.4.2
opencv-python==4.13.0.92
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
